Only need 16 bit integers

+++ src/entities/paddle.rs
@@ -1,12 +1,12 @@
 pub struct Paddle {
-	x: f64,
-	y: f64,
+	x: i16,
+	y: i16,
 	width: f64,
 	height: f64
 }
 
 impl Paddle {
-	fn position(&self) -> [f64; 2] {
+	fn position(&self) -> [i16; 2] {
 		[self.x, self.y]
 	}
 }
@@ -17,13 +17,13 @@ mod tests {
 
 	#[test]
 	fn test_basic_paddle() {
-		let p1 = Paddle{x: 15.0, y: 15.0, width: 20.0, height: 20.0};
-		assert!(p1.x > 0.0);
+		let p1 = Paddle{x: 15, y: 15, width: 20.0, height: 20.0};
+		assert!(p1.x > 0);
 	}
 
 	#[test]
 	fn test_position() {
-		let p1 = Paddle{x: 15.0, y: 15.0, width: 20.0, height: 20.0};
-		assert_eq!([15.0, 15.0], p1.position());
-	}	
+		let p1 = Paddle{x: 15, y: 15, width: 20.0, height: 20.0};
+		assert_eq!([15, 15], p1.position());
+	}
 }
